+++ picturebot/helper.py
@@ -1,4 +1,6 @@
 import os
+import sys
+import hashlib
 from dataclasses import dataclass
 import generalutils.guard as grd
 
@@ -44,3 +46,25 @@ def CreateFolder(path):
 
         # Check whether creation was successfull
         grd.Filesystem.PathExist(path)
+
+def HashFileMd5(file):
+    '''Create a md5 hash from a file
+
+    Args:
+        file (file): File where a hash is created from
+    Returns:
+        (string) Md5 hash
+    '''
+    bufSize = 32768 # Read file in 32kb chunks
+    md5 = hashlib.md5()
+
+    with open(file, 'rb') as f:
+        while True:
+            data = f.read(bufSize)
+
+            if not data:
+                break
+
+            md5.update(data)
+            
+        return md5.hexdigest()